+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 - BREAKING: `cleanup_on_failure` defaults to `false` now.
+- Add `tags` option to terminals and filtering by specific tag.
 
 ## [0.3.5] - 2025-09-18